% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linegraph.R
\name{Graph}
\alias{Graph}
\title{Line}
\usage{
Graph(data, X, Y)
}
\arguments{
\item{data}{is the data that you are working with}

\item{X}{a x value}

\item{Y}{a y value}
}
\value{
A linegraph
}
\description{
Allows the user to make a linear regression
}
