% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistogramSegregating.R
\name{Histo}
\alias{Histo}
\title{Histogram}
\usage{
Histo(data, X)
}
\arguments{
\item{data}{is the data that you are working with}

\item{X}{the column you wish to use to make histogram}
}
\value{
makes a histogram of the data you selected
}
\description{
This function will allow you to make a Histogram
}
