% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jitter.R
\name{jitter}
\alias{jitter}
\title{Taking the data a makes a jitters plot}
\usage{
jitter(data, column1, column2)
}
\arguments{
\item{data}{is the data that you are working with}

\item{column1}{First column of choice for X}

\item{column2}{Second column of choice for Y}
}
\description{
Taking the data a makes a jitters plot
}
