% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanPairwiseDif.R
\name{midmean}
\alias{midmean}
\title{Taking the data and getting a mean}
\usage{
midmean(data, column1, column2, mn)
}
\arguments{
\item{data}{is the data that you are working with}

\item{column1}{first column}

\item{column2}{second column}

\item{mn}{mean}
}
\value{
give three columns to view with mean
}
\description{
Taking the data and getting a mean
}
