% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoNa.R
\name{non}
\alias{non}
\title{Taking the data and getting rid of the NA}
\usage{
non(data)
}
\arguments{
\item{data}{is the data that you are working with}
}
\value{
will remove NAs from data and then show results of deletion
}
\description{
Taking the data and getting rid of the NA
}
