% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatter.R
\name{Scatter}
\alias{Scatter}
\title{Scatter}
\usage{
Scatter(data, X, Y)
}
\arguments{
\item{data}{is the data that you are working with}

\item{X}{x value}

\item{Y}{y value}
}
\value{
will remove NAs from data and then show results of deletion
}
\description{
This will allow the user to make a scatterplot
}
