% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection.R
\name{Trial}
\alias{Trial}
\title{Allows the user to select columns of data to use up to twelve}
\usage{
Trial(
  data,
  column1,
  column2,
  column3,
  column4,
  column5,
  column6,
  column7,
  column8,
  column9,
  column10,
  column11,
  column12
)
}
\arguments{
\item{data}{is the data that you are working with}

\item{column1-column12}{are the columns that you want to specifically use in your dataset}
}
\value{
it will return a clean dataset with the desired columns to work with
}
\description{
Allows the user to select columns of data to use up to twelve
}
